6.8

+++ homework_6/src/image.cpp
@@ -67,8 +67,8 @@ void Image::DownScale(int scale) {
 
     for (int height = 0; height < Row; height++) {
         for (int witdh = 0; witdh < Col; witdh++) {
-            posX = (int)floor(witdh * scale);
-            posY = (int)floor(height * scale);
+            posX = floor(witdh * scale);
+            posY = floor(height * scale);
             int suc = height * Col + witdh;
             int pre = posY * Image::cols_ + posX;
             imgGray.at(suc) = Image::data_.at(pre);
